Freamwork/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pcnme"
version = "0.1.0"
description = "Predictive Cloud-Native Mobile Edge (PCNME) task offloading framework"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "PCNME Contributors" }]
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "websockets>=12.0",
  "pymoo>=0.6.0",
  "torch>=2.1",
  "numpy>=1.24",
  "pandas>=2.0",
  "scipy>=1.10",
  "redis>=5.0",
  "asyncpg>=0.29",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "structlog>=24.1",
  "click>=8.1",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
osm = ["osmnx>=1.9"]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
]

[project.scripts]
pcnme = "pcnme.cli.main:cli"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["pcnme*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
